Moved oio_sys_* from cluster/lib to core/ext
(Functions to get the CPU & IO load)
