+++ core/oioext.h
@@ -69,6 +69,12 @@ gint64 oio_ext_real_time (void);
 
 gint64 oio_ext_monotonic_time (void);
 
+gdouble oio_sys_cpu_idle (void);
+
+gdouble oio_sys_io_idle (const char *vol);
+
+gdouble oio_sys_space_idle (const char *vol);
+
 #ifdef __cplusplus
 }
 #endif
